Add [Install] section to receptor@.service
When systemd is told to enable or disable a unit, it looks for an
`[Install]` section in that unit's definition to learn how to enable or
disable that unit. If the `[Install]` section is missing, then systemd
doesn't know what to do, and it ignores the enable or disable command.