BZ #1849656 - Fix you cannot use loops on 'import_tasks' statements (#20)
Turns out import_tasks imports the tasks before the actual run is
started so it cannot import tasks dynamically based on something
generated during the play. include_tasks is the right variant for
this.

The catch with include_tasks is that environment cannot be set on this
task, for this reason the contents of register_with_sources had to be
wrapped in a block to apply the environment to all tasks in it.